Make room for example project
