--- grails-app/i18n/messages.properties
@@ -1,6 +1,0 @@
-fileupload.upload.nofile = No file was sent, please select the file and try again.
-fileupload.upload.fileBiggerThanAllowed = Sent file is bigger than allowed. Max file size is {0} kb
-fileupload.upload.unauthorizedExtension = The file you sent has an unauthorized extension ({0}). Allowed extensions for this upload are {1}
-
-fileupload.download.nofile = No file was found with id {0}. Please check your link.
-fileupload.download.filenotfound = The requested file ({0}) was not found in our system.
