+++ grails-file-uploader/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+encoding//grails-app/i18n=UTF-8
+encoding//grails-app/i18n/messages.properties=UTF-8
