+++ grails-file-uploader/.settings/org.grails.ide.eclipse.core.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+org.grails.ide.eclipse.core.org.grails.ide.eclipse.core.install.name=Grails 2.2.4
+org.grails.ide.eclipse.core.use.default.install=true
