+++ grails-file-uploader/grails-app/conf/UrlMappings.groovy
