+++ grails-file-uploader/grails-app/domain/com/lucastex/grails/fileuploader/UFile.groovy
